Hygiene: rustfmt the input-desktop diagnostic, drop a useless i16 conversion
Two AC7 gate failures, both from this branch. `cargo fmt` had never been run
over the Windows-only diagnostic added earlier (the file compiles only for the
Windows target, but rustfmt is target-independent, so the branch was not
fmt-clean). Clippy's `useless_conversion` fired on the new `send_scroll`:
WHEEL_UNITS_PER_NOTCH is already i16 after the cast, so `i16::from` around it
converted a type to itself.

No behaviour change.

+++ src/autoinput.rs
@@ -249,7 +249,7 @@ fn send_scroll(input: &WakingSender<InputEvent>, x: u16, y: u16, notches: i16) -
     if input.send(InputEvent::MouseMove { x, y }).is_err() {
         return false;
     }
-    let step = i16::from(crate::input::WHEEL_UNITS_PER_NOTCH as i16) * notches.signum();
+    let step = crate::input::WHEEL_UNITS_PER_NOTCH as i16 * notches.signum();
     for _ in 0..notches.unsigned_abs() {
         std::thread::sleep(KEY_GAP);
         let sent = input
